+++ src/FactFoundry.TelemetryForge.Server/Migrations/20260628224010_AddApiEvents.Designer.cs
@@ -3,6 +3,7 @@ using System;
 using FactFoundry.TelemetryForge.Server.Data;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Infrastructure;
+using Microsoft.EntityFrameworkCore.Migrations;
 using Microsoft.EntityFrameworkCore.Storage.ValueConversion;
 using Npgsql.EntityFrameworkCore.PostgreSQL.Metadata;
 
@@ -11,9 +12,11 @@ using Npgsql.EntityFrameworkCore.PostgreSQL.Metadata;
 namespace FactFoundry.TelemetryForge.Server.Migrations
 {
     [DbContext(typeof(TelemetryForgeDbContext))]
-    partial class TelemetryForgeDbContextModelSnapshot : ModelSnapshot
+    [Migration("20260628224010_AddApiEvents")]
+    partial class AddApiEvents
     {
-        protected override void BuildModel(ModelBuilder modelBuilder)
+        /// <inheritdoc />
+        protected override void BuildTargetModel(ModelBuilder modelBuilder)
         {
 #pragma warning disable 612, 618
             modelBuilder